Updates: user can delete only his/her posts, can get all his/her posts, posts returned from newest to oldest. Updated front-end.

+++ src/main/java/com/baltictalents/Blog/controllers/BlogController.java
@@ -2,12 +2,14 @@ package com.baltictalents.Blog.controllers;
 
 import com.baltictalents.Blog.entities.Post;
 import com.baltictalents.Blog.services.PostService;
+import com.baltictalents.Blog.services.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.access.annotation.Secured;
 import org.springframework.security.access.prepost.PostAuthorize;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 
+import java.security.Principal;
 import java.util.List;
 import java.util.Optional;
 
@@ -15,27 +17,46 @@ import java.util.Optional;
 @RequestMapping("/api")
 public class BlogController {
 
-    @Autowired
-    private PostService postService;
+    private final PostService postService;
+    private final UserService userService;
+
+    public BlogController(PostService postService, UserService userService) {
+        this.postService = postService;
+        this.userService = userService;
+    }
+
 
     @GetMapping("/posts")
     public List<Post> getAllPosts() {
         return postService.getAllPosts();
     }
 
+
+    @GetMapping("/user/posts")
+    public List<Post> getAllUserPosts(Principal principal) {
+        return postService.getAllUserPosts(principal.getName());
+    }
+
+
     @GetMapping("post/{id}")
     public Optional<Post> getPost(@PathVariable Integer id) {
         return postService.getPost(id);
     }
 
+
     @PostMapping("/post")
-    public void writePost(@RequestBody Post post) {
+    public void writePost(@RequestBody Post post, Principal principal) {
+        post.setAuthor(userService.getUser(principal.getName()));
         postService.writePost(post);
     }
 
     @DeleteMapping("post/{id}")
-    public void deletePost(@PathVariable Integer id) {
-        postService.deletePost(id);
+    public void deletePost(@PathVariable Integer id, Principal principal) {
+        if(postService.getPost(id).get().getAuthor().getUsername().equals(principal.getName())){
+            postService.deletePost(id);
+        }
+
+
     }
 
 }
